Make linker a part of the publish pipeline
This change adds the linker to the publish pipeline. This makes the
"dotnet link" command obsolete, since link will be called during
"dotnet publish". The link target has been refactored into smaller
targets.

We want to introduce a target that runs the linker during the
publish pipeline. The publish pipeline has a target called
ComputeFilesToPublish, which computes the ItemGroup
ResolvedFileToPublish. To extend this target, we have two options:

1. Insert a target before ComputeFilesToPublish. Our target
would rewrite the relevant inputs (@(IntermediateAssembly) and
@(ResolvedAssembliesToPublish)). This lets ComputeFilesToPublish
be ignorant of the linker, but changes the meaning of
IntermediateAssembly and ResolvedAssembliesToPublish.

2. Insert a target after ComputeFilesToPublish. Our target would
pick apart ResolvedFileToPublish, pass some of the resolved
files to the linker, and reconstruct ResolvedFileToPublish to
include the linked assemblies. To pick apart
ComputeFilesToPublish, we break it back down into
IntermediateAssembly and ResolvedAssembliesToPublish. This
avoids changing the meaning of IntermediateAssembly and
ResolvedAssembliesToPublish, but makes it difficult for future
extensions to modify ResolvedFileToPublish (since they will need
to rely on the order in which AfterTargets run).

We'll go with option 1, for simplicity. These itemgroups should
be rewritten even during incremental build, so that publish
takes assemblies from the linked directory.

This commit also includes response to some PR feedback:

- Rename link task from "Link" to "ILLink"
- Compute excluded roots from a list of packages
- Make most targets and itemgroups private
- Turn off the size comparison report by default
- Note behavior of roots vs xml files in comment