Set nuspec version dynamically to enable include version suffix.
We used to set the version in the .nuspec file to $(VersionPrefix),
which is a static property set in the dotnet-link and ILLink.Tasks
project files. We want to use

dotnet pack --version-suffix=xxx,

which will compute $(Version) taking into account the version prefix
and the suffix passed on the command line. This computed version needs
to be passed to the nuspec, but it is evaluated after the static
NuspecProperties in the project files. Thus we add the version to the
NuspecProperties dynamically in a target that gets called during pack.