Set up MicroBuild signing
Use cecil overrides to delaysign cecil

Move strong-name props into a separate file

Sign Newtonsoft.Json.dll with 3rd-party signing cert

Because Newtonsoft.Json uses a separate signing cert from the other
dlls, it no longer makes sense to pass the authenticode cert as an
msbuild property to the signing project. To fix this, I computed the
files to sign in ILLink.Tasks.csproj (as the MicroBuild design
originally intended), and came up with some more workarounds to let
this work with a separate msbuild process on desktop.

The additional workaround is to pass the itemgroups set during publish
to the separate msbuild process as properties, and then reconstruct
the itemgroups before computing the files to sign.