Update signing logic to support multi-tfm package
This change will cause signing to happen during publish for each
target framework, instead of during pack, so that both the
netcoreapp2.0 and t he net46 dlls in the new package are signed.

This also fixes a signing issue in which real-signing was failing
because Cecil already has a strongname.