Use normal strongname signing for cecil
Remove the override file that enables delay-signing for cecil.

The microbuild instructions indicate that to use microbuild for
strongname signing, the input assemblies need to be delay-signed.
However this seems to work only when using a particular key.

Because cecil already has its own strongname key, we use that to sign
the cecil assembly instead of microbuild. This change ensures that the
cecil assemblies will be be fully signed.